modified patchDBBinaries.sh to first check if patches exist in the PATCH_DIR dir

+++ OracleDatabase/SingleInstance/extensions/patching/patchDBBinaries.sh
@@ -2,29 +2,35 @@
 # LICENSE UPL 1.0
 #
 # Copyright (c) 1982-2020 Oracle and/or its affiliates. All rights reserved.
-# 
+#
 # Since: March, 2020
 # Author: rishabh.y.gupta@oracle.com
 # Description: Applies the patches provided by the user on the oracle home.
-# 
+#
 # DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
-# 
+#
 
 PATCHES_LIST=()
 
-for patch_zip in $PATCH_DIR/*.zip; do
-    patch_no=$(echo ${patch_zip##*/} | cut -d_ -f1 | cut -dp -f2)
-    if [ $patch_no == "6880880" ]; then
-        echo "Removing directory ${ORACLE_HOME}/OPatch";
-        rm -rf ${ORACLE_HOME}/OPatch;
-        echo "Unzipping OPatch archive $patch_zip to ${ORACLE_HOME}";
-        unzip $patch_zip -d $ORACLE_HOME;
-    else
-        PATCHES_LIST+=($patch_no);
-        echo "Unzipping $patch_zip";
-        unzip $patch_zip -d $PATCH_DIR;
-    fi
-done
+ls $PATCH_DIR/*.zip;
+if [ $? -eq 0 ]; then
+    for patch_zip in $PATCH_DIR/*.zip; do
+        patch_no=$(echo ${patch_zip##*/} | cut -d_ -f1 | cut -dp -f2)
+        if [ $patch_no == "6880880" ]; then
+            echo "Removing directory ${ORACLE_HOME}/OPatch";
+            rm -rf ${ORACLE_HOME}/OPatch;
+            echo "Unzipping OPatch archive $patch_zip to ${ORACLE_HOME}";
+            unzip $patch_zip -d $ORACLE_HOME;
+        else
+            PATCHES_LIST+=($patch_no);
+            echo "Unzipping $patch_zip";
+            unzip $patch_zip -d $PATCH_DIR;
+        fi
+    done
+else
+    echo "No patches to be installed."
+    exit;
+fi
 
 export PATH=${ORACLE_HOME}/perl/bin:$PATH;
 
@@ -34,6 +40,3 @@ for patch in ${PATCHES_LIST[@]}; do
     echo "Running: $cmd";
     $cmd;
 done
-
-
-
